PointData: Move `PointData` into a separate file
